requirements.txt
pyelftools==0.29
bids-analyser==0.2.0

attrs==22.1.0
banal==1.0.6
beautifulsoup4==4.13.1
binaryornot==0.4.4
boolean.py==4.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commoncode==32.2.1
construct==2.10.68
container-inspector==31.1.0
cryptography==37.0.4
debian-inspector==31.1.0
dockerfile-parse==1.2.0
dparse2==0.7.0
extractcode==31.0.0
extractcode-7z==16.5.210531
extractcode-libarchive==3.5.1.210531
fasteners==0.17.3
fingerprints==1.0.3
ftfy==6.1.1
future==0.18.2
gemfileparser2==0.9.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
intbitset==3.1.0
isodate==0.6.1
jaraco.functools==3.5.1
javaproperties==0.8.1
Jinja2==3.1.2
jsonstreams==0.6.0
license-expression==30.4.1
lxml==5.1.0
MarkupSafe==2.1.2
more-itertools==8.13.0
normality==2.3.3
packageurl-python==0.10.0
packaging==21.3
packvers==21.5
parameter-expansion-patched==0.3.1
pdfminer.six==20220524
pefile==2022.5.30
pip-requirements-parser==32.0.1
pkginfo2==30.0.0
pluggy==1.0.0
plugincode==32.0.0
ply==3.11
publicsuffix2==2.20191221
pyahocorasick==2.0.0
pycparser==2.21
pygmars==0.9.0
Pygments==2.13.0
pymaven-patch==0.3.2
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
rdflib==6.2.0
requests==2.28.1
saneyaml==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
spdx-tools==0.8.2
text-unidecode==1.3
toml==0.10.2
typecode==30.0.1
typecode-libmagic==5.39.210531
typing-extensions==4.3.0
urllib3==1.26.11
urlpy==0.5
wcwidth==0.2.5
webencodings==0.5.1
xmltodict==0.13.0
zipp==3.8.1
